+++ src/client/ArrayGenerator.ts
