--- src/constants.ts
@@ -1,8 +1,0 @@
-
-export const blackString = "Black";
-export const whiteString = "White";
-
-export const blackLetter = blackString[0];
-export const whiteLetter = whiteString[0];
-export const emptyLetter = "E";
-export const outsideLetter = "X";
